Handle more recoil edge cases
Recoiling from a breakpoint is a giant pain in the ass. It's further
complicated by the need to support both hardware and software breakpoints and
the fact that our code for stepping and continuing is all intertwined.

However, this commit gets the code working against all six of the tests I
currently have for breakpoint handling. WOO.